adx-mcp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
azure-kusto-data==4.3.1
azure-kusto-ingest==4.3.1
azure-identity==1.15.0
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.2